requirements.txt
python-dotenv
requests
feedparser
beautifulsoup4
playwright
pytest


